Add Singularity Crafting Storage (#425)
* Add craftingStorageSingularity

* undo wrong

* AND THIS IS TO GO EVEN FURTHER BEYOND AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA

+++ src/main/java/appeng/tile/crafting/TileCraftingTile.java
@@ -234,7 +234,7 @@ public class TileCraftingTile extends AENetworkTile implements IAEMultiBlock, IP
         return false;
     }
 
-    public int getStorageBytes() {
+    public long getStorageBytes() {
         return 0;
     }
